Fix bug where rainbow boxes stayed white
To fix the bug refactored code to avoid using the rgb hex string.
This function called the getRandomRGBArray, and caused enough
overhead (even if very small) to not be able to edit CSS attributes
before a new event came about if moving the cursor completely. This
resulted in a box being marked colored but not having a BG color.